feat(pubsub): subscriber before-write callback (fast-path) (#4401)
* fix(pubsub): do not overwrite targetVariableContext

targetVariableContext contains user specific data and shall not
be overwritten

* feat(pubsub): subscriber before-write callback (fast-path)

Provide an optional before-write callback for subscriber in fast-path
configuration:
If the callback is set, the application gets a callback
before memcpy to ExternalDataValue pointer.
But the callback already knows the new received data.
This can be useful when the application either has to prepare
something before write (e.g. locks, etc.) or wants to use a
special "write" function (e.g. with signaling),
to write the variable in the underlying engineering system.

* test(pubsub): subscriber before-write callback (fast-path)

For the subscriber before-write callback test we are using
the test-plugin implementation to properly simulate
server pubsub callback processing.

The test-plugin implements a fake clock. The current time
(UA_DateTime_now) always returns the same value. Therefore the
PublishedDataSet configuration version check does not work.
UA_PubSubConfigurationVersionTimeDifference() always
returns the same value.
Therefore we have to configured the PublishedDataSet before the
DataSetWriter, otherwise the DataSet information is not used, because
the DataSetMetaData version has not changed.

* fix(pubsub): mem leak at freeze with raw encoding

There was a memory leak at freeze() reader- or writergroup with raw
encoding.
The offset DataValue has not been freed.